requirements.txt
flask
flask-cors
tensorflow
numpy
pandas
scikit-learn
python-dotenv
requests